(sitka) [RT25288] Line item number fix
The line item number LIN+# should start at 1 and increase by 1 per line
item.  This is according to the standard here:
http://www.editeur.org/files/EDIfact%20eancom%20pdfs/EDIfact%20library%20supply/L4orders%20%28library%20supply%29.pdf

This fix, assumes the user is not passing in the line item number, and
instead is using the line_index to represent the line item index from
the PO.  The fix uses the line_index first for the line_number field,
then replaces line_index with the index of the loop.